Create a ECDSARewards details component
